requirements.txt
beautifulsoup4
anthropic
ttkbootstrap
python-dotenv